api_client/requirements.txt
Django>=4.2
djangorestframework>=3.15
python-dotenv>=1.0
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.2